requirements.txt
discord.py[voice]>=2.6.4
yt-dlp>=2026.2.4
spotipy>=2.25.2
PyNaCl>=1.5.0
python-dotenv>=1.2.1
lyricsgenius>=3.7.6
aiosqlite>=0.19.0
quart>=0.20.0
hypercorn>=0.17.0
